+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [1.6.3-SNAPSHOT]
 
+- Added resources helper to App class
 - Added TrayIcon support (https://github.com/edvin/tornadofx/issues/255)
 - EventBus `fire()` function is now available from the App class
 - `ComboBox.makeAutocompletable()`